Moving courseClashDetector into controllers
