+++ controllers/courseClashDetector.js
